feat: all classes turned to template classes
additional refactors and methods needed to handle all three data types

+++ Semester 3/Efficient Programming Techniques/Laboratories/lab4/src/Main.cpp
@@ -5,7 +5,7 @@ int main() {
 
     std::cout << "[DEBUG] Welcome to the Expression Tree Interface!" << std::endl;
 
-    ExpressionTreeInterface<int>::run();
+    ExpressionTreeInterface<double>::run();
 
     return 0;
 }
